Fix: `ColonyMeta`, `ColonySubscribe` after rebase changes
